continue with changing. also change to i64

+++ examples/function/function_nest.rs
@@ -1,5 +1,5 @@
 fn main() {
-    fn another_function(x: i32) {
+    fn another_function(x: i64) {
         println!("The value of x is: {}", x);
     }
     another_function(5);
